+++ src/pages/Dashboard/index.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { View } from 'react-native';
+
+const Dashboard: React.FC = () => <View />;
+
+export default Dashboard;
