+++ src/routes/App.routes.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { createStackNavigator } from '@react-navigation/stack';
+
+import Dashboard from '../pages/Dashboard';
+
+const Auth = createStackNavigator();
+
+const AppRoutes: React.FC = () => (
+  <Auth.Navigator
+    screenOptions={{
+      // headerShown: false,
+      cardStyle: { backgroundColor: '#312e38' },
+    }}
+  >
+    <Auth.Screen name="Dashboard" component={Dashboard} />
+  </Auth.Navigator>
+);
+
+export default AppRoutes;
