+++ src/routes/index.tsx
@@ -1,21 +1,12 @@
 import React from 'react';
-import { createStackNavigator } from '@react-navigation/stack';
 
-import SignUp from '../pages/SignUp';
-import SignIn from '../pages/SignIn';
+import { useAuth } from '../hooks/auth';
+import AuthRoutes from './auth.routes';
+import AppRoutes from './App.routes';
 
-const Auth = createStackNavigator();
+const Routes: React.FC = () => {
+  const { user } = useAuth();
+  return user ? <AppRoutes /> : <AuthRoutes />;
+};
 
-const AuthRoutes: React.FC = () => (
-  <Auth.Navigator
-    screenOptions={{
-      headerShown: false,
-      cardStyle: { backgroundColor: '#312e38' },
-    }}
-  >
-    <Auth.Screen name="SignIn" component={SignIn} />
-    <Auth.Screen name="SignUp" component={SignUp} />
-  </Auth.Navigator>
-);
-
-export default AuthRoutes;
+export default Routes;
